Rename the `http` module to `helpers` to prevent conflicts with the built-in Python http library (fixes #1323)
